move ThemeProvider
