+++ frontends/ol-design/src/index.ts
@@ -83,5 +83,6 @@ export * from "./components/ButtonLink"
 export * from "./components/ChipLink"
 export * from "./components/RoutedDrawer"
 export * from "./components/SimpleMenu"
+export * from "./components/ThemeProvider"
 export * from "./components/LoadingSpinner"
 export * from "./hooks/useBreakpoint"
